speech-to-text/pyproject.toml
[project]
name = "speech-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt",
    "vosk",
    "numpy",
    "argparse",
]
